Split out S3 client into its own module and misc code cleanup
