+++ .gitignore
@@ -1,2 +1,4 @@
 /target
 .venv
+src/*.save
+src/bin/*.save
