+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "dlio_s3_rust"
-version = "0.2.0"
+version = "0.2.1"
 edition = "2024"
 
 [lib]
